MH-9720 -- fixed potential NullPointerException when dealing with non-existing vs. empty directories.
git-svn-id: https://opencast.jira.com/svn/MH/trunk@14855 e6edf6fb-f266-4316-afb4-e53d95876a76

+++ modules/matterhorn-distribution-service-streaming/src/main/java/org/opencastproject/distribution/streaming/StreamingDistributionService.java
@@ -314,7 +314,7 @@ public class StreamingDistributionService extends AbstractJobProducer implements
 
       // Try to remove the file and - if possible - the parent folder
       FileUtils.forceDelete(elementFile);
-      if (mediapackageDir.list().length == 0) {
+      if (mediapackageDir.isDirectory() && mediapackageDir.list().length == 0) {
         FileSupport.delete(mediapackageDir.getParentFile());
       }
